home(zh): add dedicated Chinese homepage
Note that navbar.html uses `relLangURL` on logo and all menu links. However, we only want to use `relLangURL` on ZH logo (to go to ZH home) and `relURL` on all other menu links in order to link to the pages on the English site. In order to achieve this, we prefix ZH menu links with `../`.